moved all rust api into mothra_api and made conditional compilation
