+++ core/Cargo-rust.toml
@@ -35,3 +35,6 @@ bytes = "0.4.12"
 env_logger = "0.6.0"
 tokio-stdin-stdout = "0.1"
 
+[features]
+default = []
+capi = []
