+++ core/src/c_api.rs
@@ -1,3 +1,4 @@
+/*
 extern crate getopts;
 use std::sync::mpsc::{channel,Sender,Receiver};
 use std::{thread, time};
@@ -18,37 +19,15 @@ use crate::mothra_api::{config,start};
 use crate::api::{send_gossip,send_rpc_request,send_rpc_response,GLOBAL};
 use crate::{get_tx,set_tx};
 
-type discovered_peer_type = unsafe extern "C" fn(peer_c_uchar: *const c_uchar, peer_length: i16);
-type receive_gossip_type = unsafe extern "C" fn(topic_c_uchar: *const c_uchar, topic_length: i16, data_c_uchar: *mut c_uchar, data_length: i16);
-type receive_rpc_type =  unsafe extern "C" fn(method_c_uchar: *const c_uchar, method_length: i16, req_resp: i16, peer_c_uchar: *const c_uchar, peer_length: i16, data_c_uchar: *mut c_uchar, data_length: i16);
-static mut s_discovered_peer_ptr: Option<discovered_peer_type> = None;
-static mut s_receive_gossip_ptr: Option<receive_gossip_type> = None;
-static mut s_receive_rpc_ptr: Option<receive_rpc_type> = None;
-
 #[no_mangle]
 pub unsafe extern "C" fn libp2p_register_handlers(
         discovered_peer_ptr: unsafe extern "C" fn(peer_c_uchar: *const c_uchar, peer_length: i16),
         receive_gossip_ptr: unsafe extern "C" fn(topic_c_uchar: *const c_uchar, topic_length: i16, data_c_uchar: *mut c_uchar, data_length: i16), 
         receive_rpc_ptr: unsafe extern "C" fn(method_c_uchar: *const c_uchar, method_length: i16, req_resp: i16, peer_c_uchar: *const c_uchar, peer_length: i16, data_c_uchar: *mut c_uchar, data_length: i16)
     ) {
-        s_discovered_peer_ptr = Some(discovered_peer_ptr);
-        s_receive_gossip_ptr = Some(receive_gossip_ptr);
-        s_receive_rpc_ptr = Some(receive_rpc_ptr);
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn discovered_peer (peer_c_uchar: *const c_uchar, peer_length: i16) {
-    s_discovered_peer_ptr.unwrap()(peer_c_uchar, peer_length);
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn receive_gossip (topic_c_uchar: *const c_uchar, topic_length: i16, data_c_uchar: *mut c_uchar, data_length: i16) {
-    s_receive_gossip_ptr.unwrap()(topic_c_uchar, topic_length, data_c_uchar, data_length);
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn receive_rpc (method_c_uchar: *const c_uchar, method_length: i16, req_resp: i16, peer_c_uchar: *const c_uchar, peer_length: i16, data_c_uchar: *mut c_uchar, data_length: i16) {
-    s_receive_rpc_ptr.unwrap()(method_c_uchar, method_length, req_resp, peer_c_uchar, peer_length, data_c_uchar, data_length);
+        crate::mothra_api::api::s_discovered_peer_ptr = Some(discovered_peer_ptr);
+        crate::mothra_api::api::s_receive_gossip_ptr = Some(receive_gossip_ptr);
+        crate::mothra_api::api::s_receive_rpc_ptr = Some(receive_rpc_ptr);
 }
 
 #[no_mangle]
@@ -83,7 +62,7 @@ pub extern "C" fn libp2p_start(args_c_char: *mut *mut c_char, length: isize) {
 
     set_tx!(&RefCell::new(Some(tx1)));
 
-    ///Listen for messages rcvd from the network
+    //Listen for messages rcvd from the network
     thread::spawn(move || {
         loop{
             match rx2.recv(){
@@ -149,3 +128,4 @@ pub extern "C" fn libp2p_send_rpc_response(method_c_uchar: *mut c_uchar, method_
     let mut data = unsafe { std::slice::from_raw_parts_mut(data_c_uchar, data_length).to_vec() };
     send_rpc_response(method, peer, data);
 }
+*/
