+++ core/src/lib.rs
@@ -1,5 +1,3 @@
 pub mod mothra_api;
-pub mod api;
-pub mod c_api;
 
-pub use api::{register_handlers,network_start,send_gossip};
+pub use mothra_api::api::{register_handlers,network_start,send_gossip};
